refactor(omokForbiddenRule): 금수룰 분리 및 인터페이스 상속
